smell-one-step-1
Created new class called WebFilter which will house the filtering
functionality of the WebData class. I also made a copy of the code in
WebData which will be moved to WebFilter and pasted it in WebFilter.
This allows the existing code to still function and for me to confirm
all desired code has been moved.