smell-one-step-6
In this last step, I browsed over the WebData class removing imports or
references which were not required. I also fixed a PEP8 naming issue in
a function in the newly created WebFilter class. Finally I re-ran all
the tests to make sure they all pass. This is the final step in
refactoring the bad smell. Not only did it shift some of the
functionality over from WebData to WebFilter helping to adhere to the
single responsibility principle it also reduced the size of some of the
functions in the Webdata Class. From evaluating this refactoring
process, no new bad smells have been introduced in the process.