smell-three-step-5
Modified both WebFilter and WebRequest constructors to accept a
data_validator parameter and store a reference to it as validator. Then
each time a function was called to the previously inherited functions,
self.validator was added to the front of the call. Also the class was
modified to no more inherit from OptionFilter.