smell-three-step-9
Created objects as attributes in Command class and passed references of
these objects to the classes which now require them. These are some of
the ones which previously inherited from OptionFilter. Ready to run
some tests but ran into issues in the CommandFilter class where the
functions were using getattr() by referencing self which now does not
work as the function is no longer inherited.