smell-three-step-10
This step is a continuation of the last step. CommandFilter’s functions
which were previously inherited from OptionFilter were referencing self
which now did not work and caused errors. To solve this, the classes
which functions were calling these were modified to pass themselves as
a parameter. The receiving functions were then also modified to use the
parameter which was passed instead of using self.
All tests were re-run, looking good :). Now for cleanup of unused code.